+++ src/Raven.CodeAnalysis/Compilation.IncrementalStateTransfer.cs
@@ -15,6 +15,7 @@ public partial class Compilation
         ArgumentNullException.ThrowIfNull(previousCompilation);
 
         if (Options.RuntimeUnitContract != previousCompilation.Options.RuntimeUnitContract
+            || Options.RuntimeTypeOfContract != previousCompilation.Options.RuntimeTypeOfContract
             || Options.MetadataImportOptions != previousCompilation.Options.MetadataImportOptions
             || Options.RuntimeIterationContract != previousCompilation.Options.RuntimeIterationContract
             || Options.RuntimePropagationContract != previousCompilation.Options.RuntimePropagationContract